feat: show inline market quote when placing trade from launcher (ARTP-1405)
