+++ src/client/src/apps/SimpleLauncher/spotlight/components/InlineQuote.tsx
@@ -1,6 +1,4 @@
-import React, { FC, useEffect, useState } from 'react'
-import { usePriceService } from './usePriceService'
-import { SpotPriceTick } from 'apps/MainRoute'
+import React, { FC } from 'react'
 import { InlineIntent } from './styles'
 import { DateTime } from 'luxon'
 import { ResultsTable, ResultsTableRow, LoadingRow } from './resultsTable'
@@ -8,31 +6,14 @@ import { MovementIcon } from '../../icons'
 import { defaultColDefs } from './utils'
 import { showCurrencyPair } from 'rt-interop/intents'
 import { usePlatform } from 'rt-platforms'
+import { useSpotPrice } from './useSpotPrice'
 interface InlineQuoteProps {
   currencyPair: string
 }
 
 export const InlineQuote: FC<InlineQuoteProps> = ({ currencyPair }) => {
-  const [quote, setQuote] = useState<SpotPriceTick>()
-  const priceService = usePriceService()
+  const quote = useSpotPrice(currencyPair)
   const platform = usePlatform()
-
-  useEffect(() => {
-    if (!priceService) {
-      return
-    }
-    const subscription = priceService
-      .getSpotPriceStream({ symbol: currencyPair })
-      .subscribe(result => {
-        setQuote(result)
-      }, console.error)
-
-    return () => {
-      if (subscription) {
-        subscription.unsubscribe()
-      }
-    }
-  }, [priceService, currencyPair])
 
   const row =
     quote && quote.symbol
